creo un formAll.js para las funcionalidades que comparte todo el formulario
