+++ src/js/form-comparte.js
@@ -12,11 +12,3 @@ const legendShare = document.querySelector(".js-legend-share");
 const formShare = document.querySelector(".js-form-share");
 const arrowShare = document.querySelector(".js-icon-arrow-share");
 
-legendShare.addEventListener("click", dropDown);
-
-function dropDown() {
-  formShare.classList.toggle("collapse");
-
-  arrowShare.classList.toggle("rotate");
-  console.log("clase rotate");
-}
