+++ src/js/main.js
@@ -6,5 +6,6 @@ const formContainer = document.querySelector(".js-div-container__form");
 import "./form-design.js";
 import "./form-rellena.js";
 import "./form-comparte.js";
+import "./formAll.js";
 import "./get-avatar.js";
 import "./card.js";
